46.The Lifecycle of a Message & its Stamps Adding the BusNameStamp
But... the BusNameStamp is one that you might want to add. Sure, Messenger used the correct bus in this case by accident, but we can be more explicit!

Head into ExternalJsonMessengerSerializer. Change this to $envelope = new Envelope() and, at the bottom, return . Add the stamp with $envelope = $envelope->with() - this is how you add a stamp - new BusNameStamp().

Then... hmm... because our transport & serializer only handle this one message... and because this one message is a command, we'll want to put the command bus here. Copy the command.bus bus name and paste. I'll add a comment that says that this is technically only needed if you need the message to be sent through a non-default bus.

Next, our serializer is great, but we didn't code very defensively. What would happen if the message contained invalid JSON... or was missing the emoji field? Would our app fail gracefully... or explode?