47.Graceful Failure in the Transport Serializer Our shiny new external_messages transport reads messages from this messages_from_external queue, which we're pretending is being populated by an external application. We're taking this JSON and, in ExternalJsonMessengerSerializer, decoding it, creating the LogEmoji object, putting it into an Envelope, even adding a stamp to it, and ultimately returning it, so that it can then be dispatched back through the message bus system.
Failing on Invalid JSON

This is looking great! But there are two improvements I want to make. First, we haven't been coding very defensively. For example, what if, for some reason, the message contains invalid JSON? Let's check for that: if null === $data, then throw a new MessageDecodingFailedException('Invalid JSON')

I'll show you why we're using this exact exception class in a minute. But let's try this with some invalid JSON and... see what happens. Go restart the worker so it sees our new code:
php bin/console messenger:consume -vv external_messages

Then, in the RabbitMQ manager, let's make a very annoying JSON mistake: add a comma after the last property. Publish that message! Ok, move over and... explosion!

  MessageDecodingFailedException: Invalid JSON

Oh, and interesting: this killed our worker process! Yep, if an error happens during the decoding process, the exception does kill your worker. That's not ideal... but in reality... it's not a problem. On production, you'll already be using something like supervisor that will restart the process when it dies.