47.Graceful Failure in the Transport Serializer The Magic of MessageDecodingFailedException
The same error! This is the difference between throwing a normal Exception in the serializer versus the special MessageDecodingFailedException. When you throw a MessageDecodingFailedException, your serializer is basically saying:

  Hey! Something went wrong... and I do want to throw an exception. But, I think we should discard this message from the queue: there is no point to trying it over and over again. kthxbai!

And that's super important. If we don't discard this message, each time our worker restarts, it will fail on that same message... over-and-over again... forever. Any new messages will start piling up behind it in the queue.

So let's change the Exception to MessageDecodingFailedException. Try it now:
php bin/console messenger:consume -vv external_messages
It will explode the first time... but the MessageDecodingFailedException should have removed it from the queue. When we run the worker now:

php bin/console messenger:consume -vv external_messages
Yep! The message is gone and the queue is empty.

Next, let's add one more superpower to this serializer. What if that outside system actually sends our app many different types of message - not only a message to log emojis, but maybe also messages to delete photos or cook some pizza! How can our serializer figure out which messages are which... and which message object to create?